Refactor graph export.
