+++ src/widgets/DisassemblerGraphView.cpp
@@ -5,12 +5,10 @@
 #include "core/MainWindow.h"
 #include "common/Colors.h"
 #include "common/Configuration.h"
-#include "common/CachedFontMetrics.h"
 #include "common/TempConfig.h"
 #include "common/SyntaxHighlighter.h"
 #include "common/BasicBlockHighlighter.h"
 #include "common/BasicInstructionHighlighter.h"
-#include "dialogs/MultitypeFileSaveDialog.h"
 #include "common/Helpers.h"
 
 #include <QColorDialog>
@@ -23,11 +21,8 @@
 #include <QToolTip>
 #include <QTextDocument>
 #include <QTextEdit>
-#include <QFileDialog>
-#include <QFile>
 #include <QVBoxLayout>
 #include <QRegularExpression>
-#include <QStandardPaths>
 #include <QClipboard>
 #include <QApplication>
 #include <QAction>
@@ -93,7 +88,7 @@ DisassemblerGraphView::DisassemblerGraphView(QWidget *parent, CutterSeekable *se
 
     // Export Graph menu
     actionExportGraph.setText(tr("Export Graph"));
-    connect(&actionExportGraph, SIGNAL(triggered(bool)), this, SLOT(on_actionExportGraph_triggered()));
+    connect(&actionExportGraph, &QAction::triggered, this, &DisassemblerGraphView::showExportDialog);
 
     // Context menu that applies to everything
     contextMenu->addAction(&actionExportGraph);
@@ -917,6 +912,21 @@ void DisassemblerGraphView::contextMenuEvent(QContextMenuEvent *event)
     }
 }
 
+void DisassemblerGraphView::showExportDialog()
+{
+    QString defaultName = "graph";
+    if (auto f = Core()->functionIn(currentFcnAddr)) {
+        QString functionName = f->name;
+        // don't confuse image type guessing and make c++ names somewhat usable
+        functionName.replace(QRegularExpression("[.:]"), "_");
+        functionName.remove(QRegularExpression("[^a-zA-Z0-9_].*"));
+        if (!functionName.isEmpty()) {
+            defaultName = functionName;
+        }
+    }
+    showExportGraphDialog(defaultName, "agf", currentFcnAddr);
+}
+
 void DisassemblerGraphView::blockDoubleClicked(GraphView::GraphBlock &block, QMouseEvent *event,
                                                QPoint pos)
 {
@@ -957,57 +967,6 @@ void DisassemblerGraphView::blockTransitionedTo(GraphView::GraphBlock *to)
     seekLocal(to->entry);
 }
 
-
-Q_DECLARE_METATYPE(DisassemblerGraphView::GraphExportType);
-
-void DisassemblerGraphView::on_actionExportGraph_triggered()
-{
-    QVector<MultitypeFileSaveDialog::TypeDescription> types = {
-        {tr("PNG (*.png)"), "png", QVariant::fromValue(GraphExportType::Png)},
-        {tr("JPEG (*.jpg)"), "jpg", QVariant::fromValue(GraphExportType::Jpeg)},
-        {tr("SVG (*.svg)"), "svg", QVariant::fromValue(GraphExportType::Svg)}
-    };
-    bool hasGraphviz = !QStandardPaths::findExecutable("dot").isEmpty()
-                       || !QStandardPaths::findExecutable("xdot").isEmpty();
-    if (hasGraphviz) {
-        types.append({
-            {tr("Graphviz dot (*.dot)"), "dot", QVariant::fromValue(GraphExportType::GVDot)},
-            {tr("Graphviz json (*.json)"), "json", QVariant::fromValue(GraphExportType::GVJson)},
-            {tr("Graphviz gif (*.gif)"), "gif", QVariant::fromValue(GraphExportType::GVGif)},
-            {tr("Graphviz png (*.png)"), "png", QVariant::fromValue(GraphExportType::GVPng)},
-            {tr("Graphviz jpg (*.jpg)"), "jpg", QVariant::fromValue(GraphExportType::GVJpeg)},
-            {tr("Graphviz PostScript (*.ps)"), "ps", QVariant::fromValue(GraphExportType::GVPostScript)},
-            {tr("Graphviz svg (*.svg)"), "svg", QVariant::fromValue(GraphExportType::GVSvg)}
-        });
-    }
-
-    QString defaultName = "graph";
-    if (auto f = Core()->functionIn(currentFcnAddr)) {
-        QString functionName = f->name;
-        // don't confuse image type guessing and make c++ names somewhat usable
-        functionName.replace(QRegularExpression("[.:]"), "_");
-        functionName.remove(QRegularExpression("[^a-zA-Z0-9_].*"));
-        if (!functionName.isEmpty()) {
-            defaultName = functionName;
-        }
-    }
-
-
-    MultitypeFileSaveDialog dialog(this, tr("Export Graph"));
-    dialog.setTypes(types);
-    dialog.selectFile(defaultName);
-    if (!dialog.exec())
-        return;
-
-    auto selectedType = dialog.selectedType();
-    if (!selectedType.data.canConvert<GraphExportType>()) {
-        qWarning() << "Bad selected type, should not happen.";
-        return;
-    }
-    QString filePath = dialog.selectedFiles().first();
-    exportGraph(filePath, selectedType.data.value<GraphExportType>());
-
-}
 
 void DisassemblerGraphView::onActionHighlightBITriggered()
 {
@@ -1055,61 +1014,6 @@ void DisassemblerGraphView::updateLayout()
     onSeekChanged(this->seekable->getOffset()); // try to keep the view on current block
 }
 
-void DisassemblerGraphView::exportGraph(QString filePath, GraphExportType type)
-{
-    bool graphTransparent = Config()->getBitmapTransparentState();
-    double graphScaleFactor = Config()->getBitmapExportScaleFactor();
-    switch (type) {
-    case GraphExportType::Png:
-        this->saveAsBitmap(filePath, "png", graphScaleFactor, graphTransparent);
-        break;
-    case GraphExportType::Jpeg:
-        this->saveAsBitmap(filePath, "jpg", graphScaleFactor, false);
-        break;
-    case GraphExportType::Svg:
-        this->saveAsSvg(filePath);
-        break;
-
-    case GraphExportType::GVDot: {
-        QFile file(filePath);
-        if (!file.open(QIODevice::WriteOnly | QIODevice::Text)) {
-            qWarning() << "Can't open file";
-            return;
-        }
-        QTextStream fileOut(&file);
-        fileOut << Core()->cmdRaw(QString("agfd 0x%1").arg(currentFcnAddr, 0, 16));
-    }
-    break;
-
-    case GraphExportType::GVJson:
-        exportR2GraphvizGraph(filePath, "json");
-        break;
-    case GraphExportType::GVGif:
-        exportR2GraphvizGraph(filePath, "gif");
-        break;
-    case GraphExportType::GVPng:
-        exportR2GraphvizGraph(filePath, "png");
-        break;
-    case GraphExportType::GVJpeg:
-        exportR2GraphvizGraph(filePath, "jpg");
-        break;
-    case GraphExportType::GVPostScript:
-        exportR2GraphvizGraph(filePath, "ps");
-        break;
-    case GraphExportType::GVSvg:
-        exportR2GraphvizGraph(filePath, "svg");
-        break;
-    }
-}
-
-void DisassemblerGraphView::exportR2GraphvizGraph(QString filePath, QString type)
-{
-    TempConfig tempConfig;
-    tempConfig.set("graph.gv.format", type);
-    qWarning() << Core()->cmdRawAt(QString("agfw \"%1\"")
-                                 .arg(filePath),
-                                 currentFcnAddr);
-}
 
 void DisassemblerGraphView::paintEvent(QPaintEvent *event)
 {
